vuexの型定義はNuxt.jsと相性が悪いため保留
