--- store/modules/app.ts
@@ -1,113 +1,0 @@
-/* eslint-disable no-console */
-import { createNamespacedHelpers } from 'vuex';
-import {
-  DefineActions,
-  DefineGetters,
-  DefineMutations
-} from 'vuex-type-helper';
-import { Todo } from '~/components/interface/Todo';
-
-const todoList: Todo[] = [
-  {
-    content: 'hogehoge',
-    created: '2019-04-30 15:30'
-  },
-  {
-    content: 'hogehogehoge',
-    created: '2019-04-30 15:40'
-  },
-  {
-    content: 'fugafuga',
-    created: '2019-04-30 16:00'
-  },
-  {
-    content: 'fugafugafuga',
-    created: '2019-04-30 16:10'
-  }
-];
-
-export interface State {
-  todos: Todo[];
-}
-
-export interface Getters {
-  todos: Todo[];
-}
-
-export interface Mutations {
-  insert: { content: string };
-  remove: { todo: Todo };
-}
-
-export interface Actions {
-  insertAction: { content: string };
-  removeAction: { todo: Todo };
-}
-
-// export const state: State = {
-//   todos: todoList
-// };
-export const state = () => ({
-  todos: todoList
-});
-
-export const getters: DefineGetters<Getters, State> = {
-  todos: state => state.todos
-};
-
-export const mutations: DefineMutations<Mutations, State> = {
-  insert: (state, { content }) => {
-    const date = new Date();
-    const formatedDate =
-      date.getFullYear() +
-      '-' +
-      ('00' + (date.getMonth() + 1)).slice(-2) +
-      '-' +
-      ('00' + (date.getDate() + 1)).slice(-2) +
-      '-' +
-      ('00' + (date.getHours() + 1)).slice(-2) +
-      ':' +
-      ('00' + (date.getMinutes() + 1)).slice(-2);
-    state.todos.unshift({
-      content: content,
-      created: formatedDate
-    });
-  },
-  remove: (state, { todo }) => {
-    for (let i = 0; i < state.todos.length; i++) {
-      const stateTodos = state.todos[i];
-      if (
-        stateTodos.content === todo.content &&
-        stateTodos.created === todo.created
-      ) {
-        console.log(`remove ${todo.content}`);
-        state.todos.splice(i, 1);
-        return;
-      }
-    }
-  }
-};
-
-export const actions: DefineActions<Actions, State, Mutations, Getters> = {
-  insertAction({ commit }, content) {
-    commit('insert', content);
-  },
-  removeAction({ commit }, todo) {
-    commit('remove', todo);
-  }
-};
-
-export const {
-  mapState,
-  mapGetters,
-  mapMutations,
-  mapActions
-} = createNamespacedHelpers<State, Getters, Mutations, Actions>('app');
-
-export const app = {
-  namespaced: true,
-  state,
-  getters,
-  mutations,
-  actions
-};
